@@ -12,44 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.3",
-    "@angular/common": "6.0.3",
-    "@angular/compiler": "6.0.3",
-    "@angular/core": "6.0.3",
-    "@angular/forms": "6.0.3",
-    "@angular/http": "6.0.3",
-    "@angular/platform-browser": "6.0.3",
-    "@angular/platform-browser-dynamic": "6.0.3",
-    "@angular/router": "6.0.3",
-    "@auth0/angular-jwt": "^2.0.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@auth0/angular-jwt": "^2.1.2",
     "angular2-materialize": "^15.1.10",
-    "core-js": "^2.4.1",
+    "core-js": "^3.6.5",
     "hammerjs": "^2.0.8",
-    "jquery": "^3.3.1",
+    "jquery": "^3.5.0",
     "materialize-css": "^0.100.2",
-    "rxjs": "^6.2.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.5",
+    "rxjs-compat": "^6.5.5",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "6.0.3",
-    "@angular/language-service": "6.0.3",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "^2.0.2",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.3.2",
+    "@angular-devkit/build-angular": "^0.803.26",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.10",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^13.13.2",
+    "codelyzer": "^5.2.2",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.5.3",
+    "protractor": "^5.4.4",
+    "rxjs-tslint": "^0.1.8",
     "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "2.7.2"
+    "tslint": "^5.20.1",
+    "typescript": "^3.8.3",
+    "webpack": "^4.43.0"
   }
 }
